Split form into common and custom time ranges selector
refs #33
