+++ public/css/module.less
@@ -50,6 +50,6 @@ ul.legend {
   }
 }
 
-form[name=form_timerangepicker_graphite] select {
+form[name=form_timerangepickercommon_graphite] select {
   width: 7.5em;
 }
